+++ private/enterprise/record-storage-engine/src/test/java/com/neo4j/internal/batchimport/RestartableImportIT.java
@@ -119,25 +119,22 @@ class RestartableImportIT
                 zip( fs, dbDirectory, testDirectory.directory( "snapshots" ).resolve( format( "killed-%02d.zip", restartCount ) ) );
 
                 if ( !completedOnItsOwn && !fs.fileExists( dbDirectory.resolve( FILE_NAME_STATE ) ) &&
-                        !fs.fileExists( dbDirectory.resolve( COMPLETED ) ) )
+                        fs.fileExists( dbDirectory ) && !fs.fileExists( dbDirectory.resolve( COMPLETED ) ) )
                 {
                     // This is a case which is, by all means, quite the edge case. This is state where an import started, but was killed
                     // immediately afterwards... in the middle of creating the store files. There have been attempts to solve this in the
                     // restartable importer, which works, but there's always some case somewhere else that breaks. This edge case is only
                     // visible in this test and for users it's just this thing where you'll need to clear out your store manually if this happens.
                     Path[] files = fs.listFiles( dbDirectory );
-                    if ( files != null )
+                    for ( Path file : files )
                     {
-                        for ( Path file : files )
+                        if ( fs.isDirectory( file ) )
                         {
-                            if ( fs.isDirectory( file ) )
-                            {
-                                fs.deleteRecursively( file );
-                            }
-                            else
-                            {
-                                fs.deleteFile( file );
-                            }
+                            fs.deleteRecursively( file );
+                        }
+                        else
+                        {
+                            fs.deleteFile( file );
                         }
                     }
                 }
